requirements.txt
transformers
torch
datasets[audio]
accelerate
pyaudio
scipy
gtts
pygame
google-generativeai
pymongo
numpy
pandas